Tidying PM/PR  and Readme
Various changes in PR.
Converted data access to PM record set requests instead of get field
value.
DRYed out the map reference.
Fixed the case for ZIP, which was pointing to the wrong user property.
Pushed colorbox.js into its own directory in custom (note that this can
be referenced directly as a location for the CSS in IP16 where there is
no longer a language and version number folder in the way.
Map placeholder location calculated instead of hardcoded.
Using .hide and .show instead of .css
Pushed map variables closer to the point of use