New Google Map Implementation
This new implementation adjusts the method by which the PR communicates
with the Map iframe. Rather than writing additional code to the map
iframe, postMessage is used to provide bi-directional communication,
thus permitting feedback from the map, such as when markers are clicked
or the map is moved around or the zoom level adjusted. This is easily
extended to capture and transmit any additional map events to the custom
PR.